requirements.txt
birdnetlib
numpy
scipy
soundfile
streamlit-audiorecorder
pydub
deep_translator
streamlit
tensorflow
librosa
requests
duckduckgo_search
